OO-76: moved route to concrete logo upload path
